adding url logging
adding url logging functionality using sqlite3. the logger is started
in a separate thread so the real bot isn't disturbed. added the url
logger as a separate class in a separate file. the class is named
URLlogger. the database name is hard-wired to url.db.

added another class that is called PrintHTML which creates a very
simple html page based on the latest entries in the database.

there is a helper script create_db.py as well to do an initial create
of the database.

Signed-off-by: anton <haglundanton+github@gmail.com>